Move mmagent to contrib
